+++ app/src/main/java/com/example/weatherapp/data/db/entity/WindCurrentWeather.java
@@ -2,19 +2,19 @@ package com.example.weatherapp.data.db.entity;
 
 
 public class WindCurrentWeather {
-    public int speed;
-    public int deg;
+    public float speed;
+    public float deg;
 
-    public WindCurrentWeather(int speed, int deg) {
+    public WindCurrentWeather(float speed, float deg) {
         this.speed = speed;
         this.deg = deg;
     }
 
-    public int getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 }
